implement the room unlocked event handling & persisting on the server side. Handle authentication of STOMP/WS messages.

+++ src/main/java/de/tim_greller/susserver/events/RoomUnlockedEvent.java
@@ -12,5 +12,5 @@ import lombok.extern.jackson.Jacksonized;
 @Jacksonized
 @Builder
 public class RoomUnlockedEvent extends Event {
-    private final long roomId;
+    private final int roomId;
 }
